+++ services/blockchain-connector/shared/sdk/decoder.js
@@ -13,6 +13,8 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
+const BluebirdPromise = require('bluebird');
+
 const { codec } = require('@liskhq/lisk-codec');
 const { hash } = require('@liskhq/lisk-cryptography');
 
@@ -24,6 +26,8 @@ const {
 	getTransactionSchema,
 	getTransactionParamsSchema,
 } = require('./schema');
+
+const { parseToJSONCompatObj } = require('../parser');
 
 const decodeAccount = async (encodedAccount) => {
 	const accountSchema = await getAccountSchema();
@@ -79,8 +83,55 @@ const decodeBlock = async (encodedBlock) => {
 	return decodedBlock;
 };
 
+const decodeResponse = async (action, response) => {
+	if (['app_getBlockByHeight', 'app_getBlockByID', 'app_getLastBlock'].includes(action)) {
+		const decodedBlock = await decodeBlock(response);
+		return parseToJSONCompatObj(decodedBlock);
+	}
+
+	if (['app_getBlocksByHeightBetween', 'app_getBlocksByIDs'].includes(action)) {
+		return BluebirdPromise.map(
+			response,
+			async block => {
+				const decodedBlock = await decodeBlock(block);
+				return parseToJSONCompatObj(decodedBlock);
+			}, { concurrency: response.length });
+	}
+
+	if (['app_getTransactionByID'].includes(action)) {
+		const decodedTransaction = await decodeTransaction(response);
+		return parseToJSONCompatObj(decodedTransaction);
+	}
+
+	if (['getTransactionsByIDs', 'getTransactionsFromPool'].includes(action)) {
+		return BluebirdPromise.map(
+			response,
+			async transaction => {
+				const decodedTransaction = await decodeTransaction(transaction);
+				return parseToJSONCompatObj(decodedTransaction);
+			}, { concurrency: response.length });
+	}
+	return response;
+};
+
+const decodeEventPayload = async (eventName, payload) => {
+	if (['app_newBlock', 'app_deleteBlock', 'app_chainForked'].includes(eventName)) {
+		const decodedBlock = await decodeBlock(payload.block);
+		return parseToJSONCompatObj(decodedBlock);
+	}
+
+	if (eventName === 'app_newTransaction') {
+		const decodedTransaction = await decodeTransaction(payload.transaction);
+		return parseToJSONCompatObj(decodedTransaction);
+	}
+
+	return payload;
+};
+
 module.exports = {
 	decodeAccount,
 	decodeBlock,
 	decodeTransaction,
+	decodeResponse,
+	decodeEventPayload,
 };
